+++ src/alm.cpp
@@ -205,7 +205,7 @@ void ALM::set_sparse_mode(const int sparse_mode) const // SPARSE
 //}
 
 void ALM::define(const int maxorder,
-                 const unsigned int nkd,
+                 const size_t nkd,
                  const int *nbody_include,
                  const double * const * const *cutoff_radii)
 {
